Working part two
Had to look this one up, but totally makes sense when I think about it!
Using the extended GCD formula is a bit overkill for what is needed
here.
