+++ 2020/13/part-two.js
@@ -1,1 +1,6 @@
 const { input } = require('./input');
+const { findWhenBusesAlign } = require('./bus');
+const { bus_ids } = input;
+
+
+console.log(findWhenBusesAlign(bus_ids));
